+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added checks to handle cases where no data is available for the specified combination of parameters in `fingertips_data()`.
 * When no data is available and `rank = TRUE`, the function now returns an empty data frame with `Polarity`, `Rank`, and `AreaValuesCount` columns set to `NA`.
 * Improved error message to inform users that no data is available and an empty data frame is returned.
+* Updated examples. 
 
 # fingertipsR 1.0.13 (2025-05-29)